@@ -1,23 +1,31 @@
-module.exports = {
-  parserOptions: {
-    sourceType: "script",
-    ecmaVersion: "latest",
-    ecmaFeatures: {
-      impliedStrict: true,
-      jsx: true,
+import globals from "globals";
+import eslintConfigStyle from "@cinna-dev/eslint-config-style";
+import eslintConfigError from "@cinna-dev/eslint-config-error";
+import eslintConfigSuggestions from "@cinna-dev/eslint-config-suggestions";
+import eslintConfigTs from "@cinna-dev/eslint-config-ts";
+import eslintConfigReact from "@cinna-dev/eslint-config-react";
+import eslintConfigImport from "@cinna-dev/eslint-config-import";
+
+/** @type { import("eslint").Linter.FlatConfig[] } */
+export default [
+    {
+        languageOptions: {
+            parserOptions: {
+                globals: globals.browser,
+                tsconfigRootDir: import.meta.dirname,
+                sourceType: 'script',
+                ecmaVersion: 'latest',
+                ecmaFeatures: {
+                    impliedStrict: true,
+                    jsx: true
+                }
+            }
+        }
     },
-  },
-  env: {
-    browser: true,
-    node: true,
-    es2022: true,
-  },
-  extends: [
-    "@cinna-dev/eslint-config-style",
-    "@cinna-dev/eslint-config-error",
-    "@cinna-dev/eslint-config-suggestions",
-    "@cinna-dev/eslint-config-ts",
-    "@cinna-dev/eslint-config-react",
-    "@cinna-dev/eslint-config-import",
-  ],
-};
+    ...eslintConfigStyle,
+    ...eslintConfigError,
+    ...eslintConfigSuggestions,
+    ...eslintConfigTs,
+    ...eslintConfigReact,
+    ...eslintConfigImport,
+];
